+++ ricecooker/utils/downloader.py
@@ -1,3 +1,4 @@
+import concurrent.futures
 import copy
 import mimetypes
 import os
@@ -18,7 +19,8 @@ from ricecooker.utils.caching import CacheForeverHeuristic, FileCache, CacheCont
 DOWNLOAD_SESSION = requests.Session()                          # Session for downloading content from urls
 DOWNLOAD_SESSION.mount('https://', requests.adapters.HTTPAdapter(max_retries=3))
 DOWNLOAD_SESSION.mount('file://', FileAdapter())
-cache = FileCache('.webcache')
+# use_dir_lock works with all filesystems and OSes
+cache = FileCache('.webcache', use_dir_lock=True)
 forever_adapter= CacheControlAdapter(heuristic=CacheForeverHeuristic(), cache=cache)
 
 DOWNLOAD_SESSION.mount('http://', forever_adapter)
@@ -334,6 +336,52 @@ def download_static_assets(doc, destination, base_url,
 
     return doc
 
+
+def get_archive_filename(url, page_domain=None, page_url=None, download_root=None, urls_to_replace=None):
+    file_url_parsed = urlparse(url)
+    page_url_parsed = None
+    LOGGER.info("page_domain = {}, page_url = {}".format(page_domain, page_url))
+    if page_url:
+        page_url_parsed = urlparse(page_url)
+        if not page_domain and page_url_parsed.netloc:
+            page_domain = page_url_parsed.netloc
+
+    no_scheme_url = url
+    if file_url_parsed.scheme != '':
+        no_scheme_url = url.replace(file_url_parsed.scheme + '://', '')
+    rel_path = file_url_parsed.path.replace('%', '_')
+    domain = file_url_parsed.netloc.replace(':', '_')
+    if not domain and page_domain:
+        domain = page_domain
+        if rel_path.startswith('/') and not url.startswith('/'):
+            # urlparse assumes links that are relative are relative to the domain root, which is not
+            # a safe assumption. Just use the original passed in URL for further processing.
+            rel_path = url
+    if rel_path.startswith('/'):
+        rel_path = rel_path[1:]
+    else:
+        # it is relative to the current subdir, not the root of the domain
+        if page_url_parsed.path.endswith('/'):
+            rel_path = page_url_parsed.path + rel_path
+        else:
+            rel_path = os.path.dirname(page_url_parsed.path) + '/' + rel_path
+    url_local_dir = os.path.join(domain, rel_path)
+    assert domain in url_local_dir
+
+    LOGGER.warning("rel_path = {}, parsed_url_path = {}".format(rel_path, page_url_parsed.path))
+    LOGGER.warning("domain = {}, local_dir_name = {}".format(domain, url_local_dir))
+
+    _path, ext = os.path.splitext(url_local_dir)
+    local_dir_name = os.path.dirname(url_local_dir)
+    if ext != '':
+        local_dir_name = os.path.dirname(url_local_dir)
+    if local_dir_name != url_local_dir and urls_to_replace is not None:
+        full_dir = os.path.join(download_root, local_dir_name)
+        os.makedirs(full_dir, exist_ok=True)
+        urls_to_replace[url] = no_scheme_url
+    return url_local_dir
+
+
 def archive_page(url, download_root, link_policy=None, run_js=False):
     """
     Download fully rendered page and all related assets into ricecooker's site archive format.
@@ -381,46 +429,12 @@ def archive_page(url, download_root, link_policy=None, run_js=False):
     # get related assets
     base_url = url[:url.rfind('/')]
     urls_to_replace = {}
-    def html5_derive_filename(url):
-        LOGGER.warning("url = {}".format(url))
-        file_url_parsed = urlparse(url)
-
-        no_scheme_url = url
-        if file_url_parsed.scheme != '':
-            no_scheme_url = url.replace(file_url_parsed.scheme + '://', '')
-        rel_path = file_url_parsed.path.replace('%', '_')
-        domain = file_url_parsed.netloc.replace(':', '_')
-        if not domain:
-            domain = page_domain
-            if rel_path.startswith('/') and not url.startswith('/'):
-                # urlparse assumes links that are relative are relative to the domain root, which is not
-                # a safe assumption. Just use the original passed in URL for further processing.
-                rel_path = url
-        if rel_path.startswith('/'):
-            rel_path = rel_path[1:]
-        else:
-            # it is relative to the current subdir, not the root of the domain
-            if parsed_url.path.endswith('/'):
-                rel_path = parsed_url.path + rel_path
-            else:
-                rel_path = os.path.dirname(parsed_url.path) + '/' + rel_path
-        url_local_dir = os.path.join(domain, rel_path)
-        assert domain in url_local_dir
-        LOGGER.warning("rel_path = {}, parsed_url_path = {}".format(rel_path, parsed_url.path))
-        LOGGER.warning("local_dir_name = {}".format(url_local_dir))
-
-        _path, ext = os.path.splitext(url_local_dir)
-        local_dir_name = url_local_dir
-        if ext != '':
-            local_dir_name = os.path.dirname(url_local_dir)
-        if local_dir_name != url_local_dir:
-            full_dir = os.path.join(download_root, local_dir_name)
-            os.makedirs(full_dir, exist_ok=True)
-            urls_to_replace[url] = no_scheme_url
-        return url_local_dir
 
     if content:
         LOGGER.warning("Downloading assets for {}".format(url))
+        page_url = url
+        def html5_derive_filename(url):
+            return get_archive_filename(url, page_domain, page_url, download_root, urls_to_replace)
         download_static_assets(content, download_root, base_url, derive_filename=html5_derive_filename,
                                link_policy=link_policy)
 
@@ -477,3 +491,38 @@ def archive_page(url, download_root, link_policy=None, run_js=False):
 
 def _is_blacklisted(url, url_blacklist):
     return any((item in url.lower()) for item in url_blacklist)
+
+
+def download_in_parallel(urls, func=None, max_workers=5):
+    """
+    Takes a set of URLs, and downloads them in parallel
+    :param urls: A list of URLs to download in parallel
+    :param func: A function that takes the URL as a parameter.
+                 If not specified, defaults to a session-managed
+                 requests.get function.
+    :return: A dictionary of func return values, indexed by URL
+    """
+    if func is None:
+        func = requests.get
+
+    results = {}
+    start = 0
+    end = len(urls)
+    batch_size = 100
+    with concurrent.futures.ThreadPoolExecutor(max_workers=max_workers) as executor:
+        while start < end:
+            batch = urls[start:end]
+            futures = {}
+            for url in batch:
+                futures[executor.submit(func, url)] = url
+
+            for future in concurrent.futures.as_completed(futures):
+                url = futures[future]
+                try:
+                    result = future.result()
+                    results[url] = result
+                except:
+                    raise
+            start = start + batch_size
+
+    return results